refactor : GameResult에게 MatchResult 구하는 책임을 이동하는 리팩토링
