+++ src/main/java/domain/gamer/GameResult.java
@@ -5,13 +5,34 @@ import exception.MissResultException;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
+import static java.util.stream.Collectors.collectingAndThen;
 import static java.util.stream.Collectors.toMap;
 
 public class GameResult {
     private Map<Player, MatchResult> gameResult;
 
-    public GameResult(Map<Player, MatchResult> gameResult) {
-        this.gameResult = gameResult;
+    public GameResult(Gamers gamers) {
+        this.gameResult = gamers.getPlayers().stream()
+                .collect(collectingAndThen(toMap(Player -> Player,
+                        player -> findMatchResult(player, gamers.getDealer()),
+                        (a, b) -> a),
+                        LinkedHashMap::new));
+    }
+
+    private MatchResult findMatchResult(Player player, Dealer dealer) {
+        if (player.isBlackJack() && dealer.isBlackJack()) {
+            return MatchResult.DRAW;
+        }
+
+        if (player.isBlackJack()) {
+            return MatchResult.BLACKJACK;
+        }
+
+        if (dealer.isBlackJack()) {
+            return MatchResult.LOSE;
+        }
+
+        return MatchResult.of(player.score.getScore(), dealer.score.getScore());
     }
 
     public Map<Player, Money> getPlayersTotalEarning() {
@@ -29,5 +50,9 @@ public class GameResult {
                 .map(Money::reversion)
                 .reduce(Money::sum)
                 .orElseThrow(MissResultException::new);
+    }
+
+    public Map<Player, MatchResult> getGameResult() {
+        return gameResult;
     }
 }
